# docs/swagger.yaml
basePath: /api/v1
definitions:
  operate.ModelError:
    properties:
      error_message:
        type: string
    type: object
...
  request.CreateFilm:
    properties:
      actors:
        items:
          type: integer
        type: array
      data_publish:
        example: 12.02.2023
        format: date
        type: string
      description:
        example: Futuristic film
        type: string
      name:
        example: Dune
        type: string
      rating:
        example: 9
        format: uint8
        type: integer
    type: object
...
  request.UpdateFilm:
    properties:
      actors:
        items:
          type: integer
        type: array
      data_publish:
        example: 12.02.2023
        format: date
        type: string
      description:
        example: Futuristic film
        type: string
      name:
        example: Dune
        type: string
      rating:
        example: 9
        format: uint8
        type: integer
    type: object
...
  response.Film:
    properties:
      actors:
        items:
          $ref: '#/definitions/response.FilmActors'
        type: array
      data_publish:
        example: 12.02.2023
        format: date
        type: string
      description:
        example: Futuristic film
        type: string
      id:
        example: 5
        format: uint64
        type: integer
      name:
        example: Dune
        type: string
      rating:
        example: 9
        format: uint8
        type: integer
    type: object
  response.FilmActors:
    properties:
      birthday:
        example: 12.02.2002
        format: date
        type: string
      id:
        example: 5
        format: uint64
        type: integer
      name:
        example: Тимоти Шаламе
        type: string
      sex:
        enum:
        - male
        - female
        example: male
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vet_v2002@mail.ru
    name: Artem Vetoshkin
  description: Rest API Для управления базо данных фильмов
  title: Фильмотека
  version: "1.0"
paths:
...
  /film:
    post:
      consumes:
      - application/json
      description: Добавляет фильм включая его название, описание, рейтинг, дату публикации
        и список игравших в нём актёров.
      parameters:
      - description: Информация о добавляемом фильме
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateFilm'
      produces:
      - application/json
      responses:
        "201":
          description: Фильм успешно добавлен в базу
          schema:
            $ref: '#/definitions/response.Film'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на создание фильма
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Добавление фильма.
      tags:
      - film
  /film/{film_id}:
    delete:
      description: Удаляет информацию о фильме из системы по его id.
      parameters:
      - description: Уникальный идентификатор фильма
        in: path
        name: film_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Фильм успешно удалён
        "400":
          description: В теле запросе ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на удаление фильма
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Фильм с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Удаление фильма.
      tags:
      - film
    put:
      consumes:
      - application/json
      description: Обновляет данные об фильме. Все переданные поля будут обновлены.
        Отсутствующие поля /
      parameters:
      - description: Уникальный идентификатор фильма
        in: path
        name: film_id
        required: true
        type: integer
      - description: Информация об обновлении
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateFilm'
      produces:
      - application/json
      responses:
        "200":
          description: Фильм успешно обновлен в базе
          schema:
            $ref: '#/definitions/response.Film'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на обновление фильма
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Фильм с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Обновление данных об фильме.
      tags:
      - film
  /film/list:
    get:
      description: Позволяет получить список фильмом отсортированный по определённому
        полю. А также можно делать поиска списка фильма по имени актёра или названии
        фильма. Если параметры "search_by" и "search_string" не указаны, поиск не
        производится.
      parameters:
      - default: DESC
        description: Порядок сортировки. Возможна сортировка по возрастанию 'asc'
          или по убыванию 'desc'.
        enum:
        - DESC
        - ASC
        in: query
        name: sort_order
        type: string
      - default: rating
        description: Параметр сортировки. Возможна сортировка по рейтингу 'rating',
          имени 'name' и дате публикации 'publish_date'.
        enum:
        - rating
        - name
        - publish_date
        in: query
        name: sort_by
        type: string
      - description: Параметр поиска. Возможен поиск по фрагменту имени актёра 'actor'
          или фрагменту названия фильма 'film'. Обязателен при указании параметра
          'search_name'.
        enum:
        - actor
        - film
        in: query
        name: search_by
        type: string
      - description: Фргамнет, по которому осуществляется поиск. Обязателен при указании
          параметра 'search_by'
        in: query
        name: search_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список фильмом успешно сформирован
          schema:
            items:
              $ref: '#/definitions/response.Film'
            type: array
        "400":
          description: В теле запросе ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Получение списка фильмов.
      tags:
      - film
...
schemes:
- http
securityDefinitions:
  sessionCookie:
    description: Уникальный идентификационный номер сессии
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"